feat: work with ruby 3+
Work with ruby 3.0 and 3.1.

In order to do this, we need to stop using the admantium gem, which
does not work with ruby 3+. Removing this led to the discovery that
the memoisation of value objects did not actually speed up the most
common use cases (validation a single barcode or validating a whole
bunch of different barcodes), but did lead to many objects being in
memory and not garbage-collectable. So we've stopped doing that and
now we should stop constantly growing our memory footprint.