Fix: re-expose GTIN#check_digit
PR #3 changed the internals of BarcodeValidation::GTIN to support
different GTIN formats and to allow conversion between them. While this is
a _very_ nice feature, it unfortunately made the previously public method
`GTIN#check_digit` (which was added via BarcodeValidation::Mixin) private.

In v2.2.0 we relied on this in our tests to fix the check digit of randomly
generated barcodes. v2.3.0 makes both `check_digit` and `expected_check_digit`
private, so it does not publicly expose any of the tools we'd need for
our purposes.

The simplest fix seemed to make `check_digit` public, which restores the
API that was available in v2.2.0. I've added a spec that calls the
method and its validation as a means to ensure it remains public.

Ref: https://github.com/marketplacer/barcodevalidation/pull/3