Support custom GTINs
This refactors GTIN handler class lookup so it can be extended without
monkeypatching. Developers can add/remove classes in
`BarcodeValidation::GTIN.gtin_classes` to control which class wraps a
GTIN. The use case I need this for is internall dummy GTINs that are
longer than any of the standard ones.

The Readme is updated to describe the feature.