Implemented GTIN::Base.prioritize_before
This is part of #29, which makes it easier to implement custom GTINs as
a user of this library. The first implementation in a0d0e9b worked, but
required the user to directly manipulate registered classes inside an
Array. As @beet pointed out, it's better to not expose implementation
details like that and look for a more elegant interface.

This commit changes the implementation by replacing `gtin_classes` with
`prioritized_gtin_classes` and adding a note that this is a private API.
Any subclass of `GTIN::Base` will be automatically appended to this list.

The user gets two tools directly inside `GTIN::Base`:

- `prioritize_before (other_class)`: moves this class to directly in
  front of the other class. This allows this class to be asked if it
  `handles?` a GTIN before the other class.
- `abstract_class`: lets a class remove itself from the list of classes
  that are asked if they handle a GTIN.

There are tests for both the implementation details and functionality.